+++ trueseeing/app/_dummy.py
@@ -14,3 +14,21 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+This file must be kept in a python2 and python3 compatible syntax.
+"""
+
+from __future__ import print_function  # this is here for the version check to work on Python 2.
+
+import sys
+
+def invoke():
+  if sys.version_info < (3, 0):
+      print("#" * 76, file=sys.stderr)
+      print("# trueseeing requires Python 3.0 or higher!                                #", file=sys.stderr)
+      print("#" * 76, file=sys.stderr)
+      sys.exit(1)
+  else:
+    import trueseeing.app.shell
+    trueseeing.app.shell.Shell().invoke()
