+++ trueseeing/app/exploit.py
@@ -1,0 +1,117 @@
+# -*- coding: utf-8 -*-
+# Trueseeing: Non-decompiling Android application vulnerability scanner
+# Copyright (C) 2017 Takahiro Yoshimura <takahiro_y@monolithworks.co.jp>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+import glob
+import os
+import re
+import tempfile
+import logging
+import shutil
+
+import lxml.etree as ET
+
+from trueseeing.core.patch import Patch, Patcher
+from trueseeing.core.sign import SigningKey, Unsigner, Resigner
+
+log = logging.getLogger(__name__)
+
+class ExploitMode:
+  def __init__(self, files):
+    self._files = files
+
+  def invoke(self, mode):
+    for f in self._files:
+      if mode == 'resign':
+        ExploitResign(f, os.path.basename(f).replace('.apk', '-resigned.apk')).exploit()
+      elif mode == 'unsign':
+        ExploitUnsign(f, os.path.basename(f).replace('.apk', '-unsigned.apk')).exploit()
+      elif mode == 'enable-debug':
+        ExploitEnableDebug(f, os.path.basename(f).replace('.apk', '-debuggable.apk')).exploit()
+      elif mode == 'enable-backup':
+        ExploitEnableBackup(f, os.path.basename(f).replace('.apk', '-backupable.apk')).exploit()
+      elif mode == 'disable-pinning':
+        ExploitDisablePinning(f, os.path.basename(f).replace('.apk', '-unpinned.apk')).exploit()
+    return 0
+
+class ExploitUnsign:
+  def __init__(self, apk, out):
+    self._unsigner = Unsigner(apk, out)
+
+  def exploit(self):
+    self._unsigner.unsign()
+
+class ExploitResign:
+  def __init__(self, apk, out):
+    self._resigner = Resigner(apk, out)
+
+  def exploit(self):
+    self._resigner.resign()
+
+class ExploitEnableDebug(Patch):
+  def __init__(self, apk, out):
+    self._patcher = Patcher(apk, out)
+
+  def exploit(self):
+    return self._patcher.apply(self)
+
+  def apply(self, context):
+    manifest = context.parsed_manifest()
+    for e in manifest.xpath('.//application'):
+      e.attrib['{http://schemas.android.com/apk/res/android}debuggable'] = "true"
+    with open(os.path.join(context.wd, 'AndroidManifest.xml'), 'wb') as f:
+      f.write(ET.tostring(manifest))
+
+class ExploitEnableBackup(Patch):
+  def __init__(self, apk, out):
+    self._patcher = Patcher(apk, out)
+
+  def exploit(self):
+    return self._patcher.apply(self)
+
+  def apply(self, context):
+    manifest = context.parsed_manifest()
+    for e in manifest.xpath('.//application'):
+      e.attrib['{http://schemas.android.com/apk/res/android}allowBackup'] = "true"
+    with open(os.path.join(context.wd, 'AndroidManifest.xml'), 'wb') as f:
+      f.write(ET.tostring(manifest))
+
+class ExploitDisablePinning(Patch):
+  def __init__(self, apk, out):
+    self._patcher = Patcher(apk, out)
+
+  def exploit(self):
+    return self._patcher.apply(self)
+
+  def patched(self, context):
+    manifest = context.parsed_manifest()
+    for e in manifest.xpath('.//application'):
+      e.attrib['{http://schemas.android.com/apk/res/android}networkSecurityConfig'] = "@xml/network_security_config"
+    with open(os.path.join(context.wd, 'AndroidManifest.xml'), 'wb') as f:
+      f.write(ET.tostring(manifest))
+    os.makedirs(os.path.join(context.wd, 'res', 'xml'), exist_ok=True)
+    with open(os.path.join(context.wd, 'res','xml','network_security_config.xml'), 'wb') as f:
+      f.write(b'''\
+<?xml version="1.0" encoding="utf-8"?>
+<network-security-config>
+  <base-config cleartextTrafficPermitted="true">
+    <trust-anchors>
+      <certificates src="system" />
+      <certificates src="user" />
+    </trust-anchors>
+  </base-config>
+</network-security-config>
+''')
