+++ trueseeing/core/code/op.py
@@ -14,3 +14,32 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
+import attr
+
+class Token:
+  def __init__(self, t=None, v=None):
+    self.t = t
+    self.v = v
+
+  def __repr__(self):
+    return '<Token t={} v={}>'.format(self.t, self.v)
+
+  def eq(self, t, v):
+    return (self.t, self.v) == (t, v)
+
+class Op:
+  def __init__(self, t=None, v=None, p=None, id_=None):
+    self.t = t
+    self.v = v
+    self.p = p
+    self._id = id_
+
+  def __repr__(self):
+    return '<Op t={} v={}, p={}>'.format(self.t, self.v, self.p)
+
+  def eq(self, t, v):
+    return (self.t, self.v) == (t, v)
+
+  @staticmethod
+  def of_id(id_):
+    return Op(None, None, None, id_)
