+++ trueseeing/core/tools.py
@@ -15,21 +15,14 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-import os.path
-import pkg_resources
+import logging
+log = logging.getLogger(__name__)
 
-class Store:
-  def __init__(self, c):
-    self.c = c
-
-  def stage0(self):
-    with open(pkg_resources.resource_filename(__name__, os.path.join('libs', 'store.s.sql')), 'r', encoding='utf-8') as f:
-      self.c.executescript(f.read())
-
-  def stage1(self):
-    with open(pkg_resources.resource_filename(__name__, os.path.join('libs', 'store.0.sql')), 'r', encoding='utf-8') as f:
-      self.c.executescript(f.read())
-
-  def stage2(self):
-    with open(pkg_resources.resource_filename(__name__, os.path.join('libs', 'store.1.sql')), 'r', encoding='utf-8') as f:
-      self.c.executescript(f.read())
+def noneif(x, defaulter):
+  if x is not None:
+    return x
+  else:
+    if callable(defaulter):
+      return defaulter()
+    else:
+      return defaulter
