+++ trueseeing/signature/fingerprint.py
@@ -20,14 +20,12 @@
 # * Fingerprinting obfuscators
 
 import collections
-import itertools
 import os
 import re
 import logging
-from trueseeing.flow.code import InvocationPattern
-from trueseeing.flow.data import DataFlows
+from trueseeing.core.flow.code import InvocationPattern
 from trueseeing.signature.base import Detector
-from trueseeing.issue import IssueConfidence, Issue
+from trueseeing.core.issue import IssueConfidence, Issue
 
 import pkg_resources
