+++ trueseeing/signature/privacy.py
@@ -30,10 +30,10 @@ import logging
 
 import re
 
-from trueseeing.flow.code import InvocationPattern
-from trueseeing.flow.data import DataFlows
+from trueseeing.core.flow.code import InvocationPattern
+from trueseeing.core.flow.data import DataFlows
 from trueseeing.signature.base import Detector
-from trueseeing.issue import IssueConfidence, Issue
+from trueseeing.core.issue import IssueConfidence, Issue
 
 log = logging.getLogger(__name__)
