+++ trueseeing/tests/test_trueseeing.py
@@ -2,14 +2,14 @@ import sys
 import os
 import unittest
 
-import trueseeing.shell
+import trueseeing.app.shell
 
 
 class TestTrueseeing(unittest.TestCase):
   def test_trueseeing(self):
     os.chdir(os.path.dirname(__file__))
     sys.argv.append('libs/Android-InsecureBankv2/InsecureBankv2.apk')
-    trueseeing.shell.shell()
+    trueseeing.app.shell.shell()
 
     # TODO assert count of severity
